refactor: take the form structure onto the base
`Form`, `FormRow`, `FormActions`, `Field`, `Help`, `Problem` and `Refused`
move out of `ui/controls.tsx` into `ui/form.tsx`, rebuilt on Tailwind with
every prop and every rendered attribute unchanged. Thirty-four consumers
change one import line and nothing else.

They land in `apps/web/ui/` rather than `apps/web/components/ui/` because
that directory holds shadcn's own primitives, added with its CLI and named
as the registry names them. These are egma's compositions of those
primitives — a `Field` wraps whichever control it is given — so they sit
beside the product's other shared components.

Two details that are not a straight class-for-class port:

- `.form > .section:first-child { margin-top: 0 }` was a descendant rule
  reaching for a class. It is now `[&>[data-slot=section]:first-child]:mt-0`
  and `Section` carries `data-slot="section"`, so the form still knows which
  section is first without knowing what a section is styled with.
- The hint and the two sentences keep `leading-(--line-normal)` written
  *after* `text-sm`. tailwind-merge drops a `leading-*` that precedes a
  `text-*`, and the line height is 1.5 here rather than the 14px step's
  own 1.43.

`field-hint.ts` said `Field` was still a CSS Modules component. It is not,
so the paragraph now says what actually holds: the context outlived the file
it was extracted from, and neither half imports the other.
