+++ apps/web/app/device/page.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 
-import { Field } from "../../ui/controls.tsx";
+import { Field } from "../../ui/form.tsx";
 import { AuthForm, AuthShell } from "../ui.tsx";
 
 /**
