+++ apps/web/app/new-project/page.tsx
@@ -7,18 +7,15 @@ import { writeJson, type Refusal } from "../../lib/api.ts";
 import { roleOf } from "../../lib/me.ts";
 import { projectLanding } from "../../lib/project-context.ts";
 import { PROJECTS_PATH, type ProjectSettings } from "../../lib/settings.ts";
+import { Button, Section, TextArea, TextInput } from "../../ui/controls.tsx";
 import {
-  Button,
   Field,
   Form,
   FormActions,
   Help,
   Problem,
   Refused,
-  Section,
-  TextArea,
-  TextInput,
-} from "../../ui/controls.tsx";
+} from "../../ui/form.tsx";
 import {
   AppShell,
   PageBody,
