+++ apps/web/app/projects/[projectId]/agents/[agentId]/connections/new/page.tsx
@@ -26,14 +26,16 @@ import { canAuthor } from "../../../../../../../lib/roles.ts";
 import {
   Button,
   ButtonLink,
+  Select,
+  TextInput,
+} from "../../../../../../../ui/controls.tsx";
+import {
   Field,
   Form,
   FormActions,
   Help,
   Problem,
-  Select,
-  TextInput,
-} from "../../../../../../../ui/controls.tsx";
+} from "../../../../../../../ui/form.tsx";
 import { Empty, Failure, Loading, NotFound } from "../../../../../../../ui/page-state.tsx";
 import { useProjectRead } from "../../../../../../../ui/resource.ts";
 import { useUnsavedChanges } from "../../../../../../../ui/settings-read.ts";
