+++ apps/web/app/projects/[projectId]/personas/[personaId]/page.tsx
@@ -26,13 +26,8 @@ import {
 } from "../../../../../lib/personas.ts";
 import { projectPath } from "../../../../../lib/project-context.ts";
 import { canAuthor } from "../../../../../lib/roles.ts";
-import {
-  Facts,
-  Field,
-  Form,
-  FormActions,
-  Refused,
-} from "../../../../../ui/controls.tsx";
+import { Facts } from "../../../../../ui/controls.tsx";
+import { Field, Form, FormActions, Refused } from "../../../../../ui/form.tsx";
 import { Dialog } from "../../../../../ui/dialog.tsx";
 import {
   Failure,
