+++ apps/web/app/projects/[projectId]/runs/[runId]/simulations/[simulationId]/page.tsx
@@ -17,11 +17,8 @@ import {
   type SimulationEvidence,
 } from "../../../../../../../lib/simulations.ts";
 import { Button } from "@/components/ui/button";
-import {
-  Actions,
-  Problem,
-  Refused,
-} from "../../../../../../../ui/controls.tsx";
+import { Actions } from "../../../../../../../ui/controls.tsx";
+import { Problem, Refused } from "../../../../../../../ui/form.tsx";
 import { Dialog } from "../../../../../../../ui/dialog.tsx";
 import {
   Failure,
