+++ apps/web/app/projects/[projectId]/settings/page.tsx
@@ -12,6 +12,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 
+import { Section, TextArea } from "../../../../ui/controls.tsx";
 import {
   Field,
   Form,
@@ -19,9 +20,7 @@ import {
   Help,
   Problem,
   Refused,
-  Section,
-  TextArea,
-} from "../../../../ui/controls.tsx";
+} from "../../../../ui/form.tsx";
 import { Failure, Loading, NotFound } from "../../../../ui/page-state.tsx";
 import { SettingsLayout } from "../../../../ui/settings-nav.tsx";
 import {
