+++ apps/web/app/projects/[projectId]/tests/[testId]/page.tsx
@@ -23,13 +23,8 @@ import {
   type TestVersionPage,
   type TestVersionRow,
 } from "../../../../../lib/tests.ts";
-import {
-  Actions,
-  Field,
-  Problem,
-  Refused,
-  Section,
-} from "../../../../../ui/controls.tsx";
+import { Actions, Section } from "../../../../../ui/controls.tsx";
+import { Field, Problem, Refused } from "../../../../../ui/form.tsx";
 import { Dialog } from "../../../../../ui/dialog.tsx";
 import { useDraftNavigation } from "../../../../../ui/draft-navigation.tsx";
 import { Failure, Loading, NotFound } from "../../../../../ui/page-state.tsx";
