+++ apps/web/test/components.test.tsx
@@ -14,7 +14,8 @@ import RunResultsAddress from "../app/runs/[runId]/page.tsx";
 import AgentsPage from "../app/projects/[projectId]/agents/page.tsx";
 import type { Me } from "../lib/me.ts";
 import { EVERY_NAVIGATION_ITEM } from "../lib/navigation.ts";
-import { Button, ButtonLink, Checkbox, Field } from "../ui/controls.tsx";
+import { Button, ButtonLink, Checkbox } from "../ui/controls.tsx";
+import { Field } from "../ui/form.tsx";
 import { DataTable, type Column } from "../ui/data-table.tsx";
 import { Dialog } from "../ui/dialog.tsx";
 import { Failure, NotFound } from "../ui/page-state.tsx";
